NIL/requirements.txt
absl-py==2.2.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
chex==0.1.89
cloudpickle==3.1.1
decorator==5.2.1
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.30
etils==1.12.2
Farama-Notifications==0.0.4
filelock==3.18.0
flax==0.10.6
fsspec==2025.5.0
gast==0.6.0
glfw==2.9.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
humanize==4.12.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
labmaze==1.0.6
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.3.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.13
packaging==25.0
pillow==11.2.1
proglog==0.1.12
protobuf==6.31.0
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-probability==0.25.0
tensorstore==0.1.75
toolz==1.0.0
torch==2.7.0
tqdm==4.67.1
treescope==0.1.9
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
